#839: The e-mail's attachments must be retrieved and stored in the DMS contextually to the operations on the database

+++ core/src/main/java/org/cmdbuild/data/converter/EmailConverter.java
@@ -100,8 +100,8 @@ public class EmailConverter implements StorableConverter<Email> {
 				.withId(emailStatusLookupId) //
 				.build());
 		email.setStatus(EmailStatus.fromName(lookup.description));
-		email.setActivityId((card.get(PROCESS_ID_ATTRIBUTE) != null) ? card
-				.get(PROCESS_ID_ATTRIBUTE, CardReference.class).getId().intValue() : null);
+		email.setActivityId((card.get(PROCESS_ID_ATTRIBUTE) != null) ? card.get(PROCESS_ID_ATTRIBUTE,
+				CardReference.class).getId() : null);
 		return email;
 	}
